+++ docs/changelog.md
@@ -7,6 +7,7 @@
 + UI adjustments.
 + Correction of zoom value.
 + Add "default zoom value" option.
++ Add "reduce" function for synthesis target path.
 
 **Development**: